refactor(Backend): :truck: rename and organize imports
