+++ backend/src/endpoint/announcement.rs
@@ -1,4 +1,3 @@
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::grpc::backend::announcement_set_server::*;
