+++ backend/src/endpoint/chat.rs
@@ -1,4 +1,3 @@
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::grpc::backend::chat_set_server::*;
