+++ backend/src/endpoint/contest.rs
@@ -1,4 +1,3 @@
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::entity::{contest::*, *};
