+++ backend/src/endpoint/education.rs
@@ -1,4 +1,3 @@
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::grpc::backend::education_set_server::*;
