+++ backend/src/endpoint/imgur.rs
@@ -1,4 +1,3 @@
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::grpc::backend::imgur_set_server::*;
