+++ backend/src/endpoint/mod.rs
@@ -10,18 +10,13 @@ pub mod testcase;
 pub mod token;
 pub mod user;
 
-pub mod tools {
+pub(self) mod tools {
+    pub use crate::entity::DebugName;
+    pub use crate::entity::*;
     pub use crate::grpc::TonicStream;
     pub use crate::init::db::DB;
     pub use crate::util::auth::Auth;
     pub use crate::util::error::Error;
-    pub use tokio::spawn;
-    pub use tokio::try_join;
-    pub use tracing::instrument;
-}
-pub mod endpoints {
-    pub use crate::entity::DebugName;
-    pub use crate::entity::*;
     pub use crate::util::pager::*;
     pub use crate::{fill_active_model, fill_exist_active_model, server::Server};
     pub use sea_orm::{
@@ -29,6 +24,9 @@ pub mod endpoints {
         PaginatorTrait, QueryFilter, QuerySelect, TransactionTrait,
     };
     pub use std::sync::Arc;
+    pub use tokio::spawn;
+    pub use tokio::try_join;
     pub use tonic::*;
+    pub use tracing::instrument;
     pub use uuid::Uuid;
 }
