+++ backend/src/endpoint/problem.rs
@@ -1,4 +1,3 @@
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::grpc::backend::problem_set_server::*;
