+++ backend/src/endpoint/testcase.rs
@@ -1,4 +1,3 @@
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::grpc::backend::testcase_set_server::*;
