+++ backend/src/endpoint/token.rs
@@ -1,6 +1,5 @@
 use std::time::Duration;
 
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::controller::token::UserPermBytes;
