+++ backend/src/endpoint/user.rs
@@ -1,4 +1,3 @@
-use super::endpoints::*;
 use super::tools::*;
 
 use crate::controller::token::UserPermBytes;
