+++ backend/src/entity/announcement.rs
@@ -1,9 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
-use sea_orm::{entity::prelude::*, FromQueryResult};
-
-use crate::endpoint::tools::Auth;
-use crate::util::error::Error;
+use super::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "announcement")]
@@ -24,7 +19,7 @@ pub struct Model {
 
 #[derive(DerivePartialModel, FromQueryResult)]
 #[sea_orm(entity = "Entity")]
-pub struct PartialAnnouncement {
+pub struct PartialModel {
     pub id: i32,
     pub title: String,
     pub contest_id: Option<i32>,
@@ -92,3 +87,37 @@ impl super::Filter for Entity {
         Err(Error::PermissionDeny("Can't write announcement"))
     }
 }
+
+pub struct ListPagerTrait;
+
+impl PagerSource for ListPagerTrait {
+    const ID: <Self::Entity as EntityTrait>::Column = Column::Id;
+
+    type Entity = Entity;
+
+    type Data = ();
+
+    const TYPE_NUMBER: u8 = 1;
+
+    fn filter(auth: &Auth, data: &Self::Data) -> Select<Self::Entity> {
+        Entity::find().filter(Column::Public.eq(true))
+    }
+}
+
+#[async_trait]
+impl PagerReflect<Entity> for PartialModel {
+    fn get_id(&self) -> i32 {
+        self.id
+    }
+
+    async fn all(query: Select<Entity>) -> Result<Vec<Self>, Error> {
+        let db = DB.get().unwrap();
+        query
+            .into_model::<Self>()
+            .all(db)
+            .await
+            .map_err(Into::<Error>::into)
+    }
+}
+
+pub type ListPager = PkPager<ListPagerTrait, PartialModel>;
