+++ backend/src/entity/chat.rs
@@ -1,9 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
-use sea_orm::entity::prelude::*;
-
-use crate::endpoint::tools::Auth;
-use crate::util::error::Error;
+use super::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "chat")]
