+++ backend/src/entity/contest.rs
@@ -1,13 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
-use sea_orm::QuerySelect;
-use sea_orm::{entity::prelude::*, sea_query::Alias, FromQueryResult, JoinType};
-
-use super::{user, user_contest};
-use crate::endpoint::tools::{Auth, DB};
-use crate::util::error::Error;
-
-use super::ParentalTrait;
+use super::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "contest")]
