+++ backend/src/entity/mod.rs
@@ -1,9 +1,10 @@
-use sea_orm::{EntityTrait, PrimaryKeyTrait, QueryFilter, Select};
+pub(self) use sea_orm::{
+    entity::prelude::*,
+    sea_query::{Alias, JoinType},
+    EntityTrait, FromQueryResult, PrimaryKeyTrait, QueryFilter, QuerySelect, Select,
+};
 
 use sea_orm::ColumnTrait;
-
-use crate::endpoint::tools::Auth;
-use crate::util::error::Error;
 
 pub mod announcement;
 pub mod chat;
@@ -15,6 +16,13 @@ pub mod test;
 pub mod token;
 pub mod user;
 pub mod user_contest;
+pub mod util;
+
+pub(self) use crate::init::db::DB;
+pub(self) use crate::util::{auth::Auth, error::Error};
+pub(self) use tonic::async_trait;
+
+pub(self) use util::paginator::{PagerReflect, PagerSource, PkPager};
 
 pub trait DebugName {
     const DEBUG_NAME: &'static str = "TEMPLATE_DEBUG_NAME";
