+++ backend/src/entity/problem.rs
@@ -1,11 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
-use sea_orm::QuerySelect;
-use sea_orm::{entity::prelude::*, sea_query::Alias, FromQueryResult, JoinType};
-
-use super::{contest, user};
-use crate::endpoint::tools::{Auth, DB};
-use crate::util::error::Error;
+use super::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[sea_orm(table_name = "problem")]
