+++ backend/src/entity/submit.rs
@@ -1,8 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
-use crate::endpoint::tools::Auth;
-use crate::util::error::Error;
-use sea_orm::entity::prelude::*;
+use super::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "submit")]
