+++ backend/src/entity/token.rs
@@ -1,7 +1,3 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
-use crate::endpoint::tools::Auth;
-use crate::util::error::Error;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
