+++ backend/src/entity/user.rs
@@ -1,9 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
-use super::{contest, problem, user_contest};
-use crate::endpoint::tools::Auth;
-use crate::util::error::Error;
-use sea_orm::entity::prelude::*;
+use super::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "user")]
