+++ backend/src/entity/user_contest.rs
@@ -1,5 +1,3 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
