+++ backend/src/entity/util/mod.rs
@@ -1,0 +1,1 @@
+pub mod paginator;
